requirements.txt
llama-cpp-python==0.1.57
numpy==1.24.3
typing_extensions==4.6.3
